requirements.txt
serpapi
firecrawl
python-dotenv
pprintpp
pytest